% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{add.pat.id}
\alias{add.pat.id}
\title{Add patient id.}
\usage{
add.pat.id(tab, id = "pat.id")
}
\arguments{
\item{tab}{table to add center}

\item{id}{variable name (default: center)}
}
\value{
Table with center add after mnppid.
}
\description{
This function adds the patient id \code{mnpaid}
to the specified table.
}
\seealso{
mnppid2mnpaid
}
