% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{check.raw.discrepancy.list}
\alias{check.raw.discrepancy.list}
\title{Check consistency of raw discrepancy list.}
\usage{
check.raw.discrepancy.list(df, check.datatypes = TRUE,
  check.content = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{data frame with the raw discrepancy list}

\item{check.datatypes}{logical whether data types shall be checked (default: TRUE)}

\item{check.content}{logical whether content shall be checked (default: TRUE)}

\item{verbose}{show content of each field being queried}
}
\description{
This function checks the consistency of the raw discrepancy list received from
the customer. Both the consistency of the data types in the data frame as well
as the content (i.e. do the table names, column names and doc IDs exist?) is checked.
}
\seealso{
generate.discrepancy.list
}
