% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cto_get_form_datasets.R
\name{cto_get_form_datasets}
\alias{cto_get_form_datasets}
\title{Download Data Attached to SurveyCTO Forms}
\usage{
cto_get_form_datasets(req, form_id, file_name = NULL)
}
\arguments{
\item{req}{A \code{httr2_request} object initialized via
\code{\link{cto_request}()}.}

\item{form_id}{Character string. The unique ID of the SurveyCTO form.}

\item{file_name}{Optional character vector. The specific name(s) of the
attached files to download (including the extension, often \code{.csv}). If \code{NULL}
(the default), all files attached to the form will be downloaded.}
}
\value{
A \code{list} of data sets. Each item in the list is named after
the original file. For example, if you download "prices.csv", you can
access it in R using \code{output$prices}. Returns
\code{NULL} invisibly if attached data files are found.
}
\description{
\code{cto_get_form_datasets()} downloads data files attached to
a specific SurveyCTO form. It downloads these data to temporary local storage,
reads them into R, and cleans up the temporary files.
}
\details{
This function scans your form for any uploaded/attached data files during
deployment, downloads these files in the background,
converts them into data frames, and cleans up after so no
cluttered files are left on your computer.
}
\examples{
\dontrun{
# Create a request object
req <- cto_request("my_server", "username", "password")

# Download all CSVs attached to a specific form
csv_data <- cto_get_form_datasets(req, "my_form_id")

# Access a specific attached file called "villages.csv"
village_df <- csv_data$villages
}
}
