% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cto_request.R
\name{cto_request}
\alias{cto_request}
\title{Create an Authenticated SurveyCTO Request Object}
\usage{
cto_request(server, username, password = NULL)
}
\arguments{
\item{server}{String. The name of your SurveyCTO server (the subdomain).
For example, if your URL is \verb{https://example.surveycto.com}, use \code{"example"}.}

\item{username}{String. The username/email for a user account.}

\item{password}{String. The password for the specified user. You should avoid
entering the password directly when calling this function, as it will be
captured by \code{.Rhistory}. Instead, leave it unset (\code{NULL}) and the default behavior
will prompt you for it interactively. If needed for automation purposes,
it's recommended to always use the environment variables.}
}
\value{
A modified and authenticated \code{httr2_request} object.
}
\description{
Initializes a \code{httr2} request object with the necessary server URL and
Basic Authentication credentials.
}
\details{
All functions in this package use \code{scto.verbose} option to control console
feedback. By default, it informs the user which process has started
and completed. You can silence this by setting \code{options(scto.verbose = FALSE)}.
}
\examples{
\dontrun{
# Direct authentication (Not Recommended)
req <- cto_request("my-org", "user@org.com", "pw")

# Input password when prompted (Interactively)
req <- cto_request("my-org", "user@org.com")

# Using environment variables for enhanced security (Recommended)
req <- cto_request("my-org", Sys.getenv("USER"), Sys.getenv("PASS"))
}
}
\seealso{
\code{\link[httr2:req_auth_basic]{httr2::req_auth_basic()}}, \code{\link[usethis:edit]{usethis::edit_r_environ()}}, \code{\link[=Sys.setenv]{Sys.setenv()}}
}
