% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{df2filename}
\alias{df2filename}
\title{Convert data frame name to file name.}
\usage{
df2filename(df)
}
\arguments{
\item{df}{data frame name or object}
}
\value{
The filename as in the SecuTrial export.
}
\description{
This function converts the internal data frame name as specified in
\code{table.list} to the filename in the export ("shorten tables names" enabled).
Either the name or the data frame directly is requeried.
}
\seealso{
table.list, df2tablename, filename2tablename
}
