% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{df2tablename}
\alias{df2tablename}
\title{Convert data frame name to original table name.}
\usage{
df2tablename(df)
}
\arguments{
\item{df}{data frame name or object}
}
\value{
The original table name as defined in SecuTrial.
}
\description{
This function converts the internal data frame name as specified in
\code{table.list} to the original table name from SecuTrial
(e.g. emnphcvtreatments).
Either the name or the data frame directly is requeried.
}
