% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{filename2tablename}
\alias{filename2tablename}
\title{Convert file name to table name.}
\usage{
filename2tablename(file.name)
}
\arguments{
\item{file.name}{filename as in the SecuTrial export}
}
\value{
The original table name as defined in SecuTrial.
}
\description{
This function converts the the filename in the export to the original table
name from SecuTrial' based on the mapping defined in \code{table.list}.
}
