% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{generate.discrepancy.list}
\alias{generate.discrepancy.list}
\title{Generate discrepancy list from data frame.}
\usage{
generate.discrepancy.list(df, df.name = deparse(substitute(df)),
  col.name, query.id = 1, query.date = format(Sys.Date(), format =
  "\%d.\%m.\%Y"), query.text = "Entry not possible.",
  query.instruction = "Please check entry.")
}
\arguments{
\item{df}{data frame typically generated by subset}

\item{df.name}{name of the original data frame as defined in \code{table.list}}

\item{col.name}{name of the column to be adressed by the query}

\item{query.id}{internal identifier of the query cycle}

\item{query.date}{date of the query (default: today)}

\item{query.text}{query text (default: Entry not possible.)}

\item{query.instruction}{instruction for the user (default: Please check entry.)}
}
\value{
The filename as in the SecuTrial export.
}
\description{
This function generates a data frame in the format required for the
"Discrepancies Transfer Module (DTM)" of SecuTrial. Input is a subset
of the original table for which for all row a query for a specified
column shall be raised.
}
\seealso{
write.discrepancy.list
}
