% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_record_table.R
\name{getDataFrameFromExportTable}
\alias{getDataFrameFromExportTable}
\title{Get dataframe from table in secuTrial export directory}
\usage{
getDataFrameFromExportTable(exportdir, table, columns = NULL,
  centres = NULL, filetype = "xls")
}
\value{
data frame from export table
}
\description{
TODO
}
