% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_record_table.R
\name{getItemsFromDRT}
\alias{getItemsFromDRT}
\title{Get items specified in secuTrial data record table}
\usage{
getItemsFromDRT(drt, tables = NULL)
}
\arguments{
\item{drt}{data record table (xls file)}

\item{tables}{the name of a specific table}
}
\value{
data frame with information on all items.
}
\description{
TODO
}
