% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_record_table.R
\name{getTablesFromDRT}
\alias{getTablesFromDRT}
\title{Get information on tables specified in data record table}
\usage{
getTablesFromDRT(drt)
}
\arguments{
\item{drt}{data record table (xls file)}
}
\value{
data frame with information on tables.
}
\description{
TODO
}
