% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load.study.options}
\alias{load.study.options}
\title{List specifying the general properties of all tables in the export.}
\usage{
load.study.options()
}
\description{
The list \code{study.options} stores all relevant technical information
in order to be able to correctly load the tables of a given study.
The list is used in the function \code{read.DB.table}.
}
\details{
\code{partial.date.handling}: By default, potential data columns containing entries that cannot be converted are
skipped (option 'skip'). With 'force.conversion' fields incompatible with the date format are converted to NA.
'fill.partial.dates' keeps the orginal column untouched and adds a new column to the data frame with the name <colname>.partial.dates.processed
in which partial are filled using \code{\link{fill.partial.date}} (e.g. Unknown.01.2013 -> 15.01.2013).
With 'fill.partial.dates.and.keep.original' partial dates are processed directly and the original data is copied to <colname>.original.
}
\seealso{
read.DB.table, convert.all.dates
}
