% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load.tables}
\alias{load.tables}
\title{Load multiple tables from an export.}
\usage{
load.tables(data.dir, tables = "all", convert.dates = FALSE,
  convert.unknown.date.to.na = FALSE, rename.headers = FALSE,
  add.pat.id = TRUE, add.center = FALSE, silent = FALSE)
}
\arguments{
\item{data.dir}{location of the export directory (or zip file) containing the csv/xls files}

\item{tables}{vector of tables to be loaded (if a \code{table.list} exists, the corresponing table names can be given)}

\item{convert.dates}{identify and convert date columns}

\item{convert.unknown.date.to.na}{convert unknown date strings (e.g 1900-01-01) to NA}

\item{rename.headers}{rename column names}

\item{add.pat.id}{adding study ID pat.id to the first column}

\item{add.center}{adding center information}

\item{silent}{hide output}
}
\value{
(Pre-processed) tables in \code{tables} as data frames
}
\description{
Load multiple tables from an export.
}
\references{
http://stackoverflow.com/questions/3640925/global-variable-in-r-function
}
\seealso{
read.DB.table, load.table.list (used in dossier-specific packages), load.study.options
}
