% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{mnpaid2mnppid}
\alias{mnpaid2mnppid}
\title{Convert patient id to internal id}
\usage{
mnpaid2mnppid(mnpaid)
}
\arguments{
\item{mnpaid}{patient id (frontend)}
}
\value{
internal id
}
\description{
This function converts the patient id \code{mnpaid}
to the internal id (\code{mnppid}) used in most tables.
It requires you to have a loaded secuTrial export in your active workspace.
}
\examples{
\dontrun{
mnpaid2mnppid(10001) # 978
}
}
\seealso{
mnppid2mnpaid, load.tables
}
