% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{mnppid2center}
\alias{mnppid2center}
\title{Convert center id to center.}
\usage{
mnppid2center(mnppid)
}
\arguments{
\item{mnppid}{internal id}
}
\value{
Center name as string.
}
\description{
This function returns the center for a given internal id \code{mnppid}.
}
