% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{mnppid2mnpaid}
\alias{mnppid2mnpaid}
\title{Convert internal id to patient id}
\usage{
mnppid2mnpaid(mnppid)
}
\arguments{
\item{mnppid}{internal id (backend)}
}
\value{
patient id
}
\description{
This function converts the internal id \code{mnppid} used in most tables
to the real patient id (\code{mnpaid}).
It requires you to have a loaded secuTrial export in your active workspace.
}
\examples{
\dontrun{
mnppid2mnpaid(978) # 10001
}
}
\seealso{
mnpaid2mnppid, load.tables
}
