% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.R
\name{move.column.after}
\alias{move.column.after}
\title{Move a column in a data frame after a specified column.}
\usage{
move.column.after(df, col.name, col.name.after)
}
\arguments{
\item{df}{data frame}

\item{col.name}{column name or vector of column names to move}

\item{col.name.after}{column name after which to place \code{col.name} ("first" to place at beginning)}
}
\value{
Data frame with shuffled columns.
}
\description{
This function moves a given column or a vector of columns (defined by column name(s))
after a specific position in the data frame (defined by column name).
}
\examples{
move.column.after(data.frame(a=NA, b=NA, c=NA, d=NA),c("c","d"),"a")
##   a  c  d  b
##1 NA NA NA NA
}
\seealso{
move.column.to.pos
}
\author{
Pascal Benkert
}
