% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.R
\name{move.column.to.pos}
\alias{move.column.to.pos}
\title{Move a column in a data frame by index.}
\usage{
move.column.to.pos(df, col.idx, new.col.idx)
}
\arguments{
\item{df}{data frame}

\item{col.idx}{old column index}

\item{new.col.idx}{new column index}
}
\value{
Data frame with shuffled columns.
}
\description{
This function moves a given column (defined by the column index)
to a specific position in the data frame.
}
\examples{
move.column.to.pos(data.frame(a=NA, b=NA, c=NA),1,3)
##   b  c  a
##1 NA NA NA
}
\seealso{
move.column.after
}
\author{
Pascal Benkert
}
