% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{process.raw.discrepancy.list}
\alias{process.raw.discrepancy.list}
\title{Add meta information to discrepancy list from customer.}
\usage{
process.raw.discrepancy.list(df, query.id = 1,
  query.date = format(Sys.Date(), format = "\%d.\%m.\%Y"),
  translate.table.name = TRUE)
}
\arguments{
\item{df}{data frame with the raw discrepancy list}

\item{query.id}{internal identifier of the query cycle}

\item{query.date}{date of the query (default: today)}

\item{translate.table.name}{translate filename to tablename (default: TRUE)}
}
\description{
This function adds the necessary meta information to the discrepancy list
needed for an import into SecuTrial using the "Discrepancies Transfer Module (DTM)".
The original column names are restored by replacing dots by underscores
in column names.
}
\seealso{
generate.discrepancy.list
}
