% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read.DB.table}
\alias{read.DB.table}
\title{Extended version of read.table.}
\usage{
read.DB.table(path, convert.dates = FALSE,
  convert.unknown.date.to.na = FALSE, rename.headers = FALSE,
  add.pat.id = TRUE, add.center = TRUE, silent = FALSE, ...)
}
\arguments{
\item{path}{location of the csv-file or unz(zip-file, table filename)}

\item{convert.dates}{flag for conversion of date columns}

\item{convert.unknown.date.to.na}{flag for conversion of unknown dates}

\item{rename.headers}{flag for renaming of headers}

\item{add.pat.id}{adding study ID pat.id to the first column}

\item{silent}{hide verbose output}

\item{...}{other options passed to \code{read.table()}}
}
\value{
Data frame of the table.
}
\description{
This function loads a database table and:
\itemize{
\item removes trailing empty columns
\item identifies and converts date columns (optionally sets "1900-01-01" to NA)
\item renames colnames (using \code{\link{new.names}})
}
}
\seealso{
load.study.options, new.names, convert.all.dates, unz
}
