% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{remove.center.tag}
\alias{remove.center.tag}
\title{Helper function to strip away center tag.}
\usage{
remove.center.tag(x)
}
\arguments{
\item{x}{center string}
}
\value{
cleaned center string
}
\description{
Example: "Universitätsspital Basel (SWISS-AF)"
}
\seealso{
mnppid2center
}
