% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secuTrial-package.R
\docType{package}
\name{secuTrial}
\alias{secuTrial}
\alias{secuTrial-package}
\title{secuTrial package}
\description{
This R package provides functions for handling data
from the clinical data management system secuTrial.
}
