% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_management.R
\name{write.discrepancy.list}
\alias{write.discrepancy.list}
\title{Write discrepancy list to csv.}
\usage{
write.discrepancy.list(df, file)
}
\arguments{
\item{df}{data frame generate by \code{link{generate.discrepancy.list}}}

\item{file}{file name for the discrepancy list csv file}
}
\value{
Discrepancy list in tab-separated csv format.
}
\description{
Simple wrapper to \code{link{write.table}} with row.names=TRUE, col.names=FALSE and sep="\\t".
The row.names build the unique identifier needed in the
discrepancy list.
}
\seealso{
generate.discrepancy.list
}
