% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{add.center}
\alias{add.center}
\title{Add center}
\usage{
add.center(tab, id = "center")
}
\arguments{
\item{tab}{table to which center column should be added to}

\item{id}{column name (default: center)}
}
\value{
Table with additional center column.
}
\description{
This function adds the center information to the specified table.
The column is added after \code{pat.id} if present and after
\code{mnppid} otherwise.
}
\seealso{
mnppid2center
}
