% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving.R
\name{move.column.after}
\alias{move.column.after}
\title{Move a column in a data frame after a specified column.}
\usage{
move.column.after(df, col.name, col.name.after)
}
\arguments{
\item{df}{data frame}

\item{col.name}{column name or vector of column names to move}

\item{col.name.after}{column name behind which to place \code{col.name} ("first" to place at beginning)}
}
\value{
Data frame with rearranged columns.
}
\description{
This function moves a given column or a vector of columns
behind a specific position in the data frame.
Columns are specified by their names.
}
\examples{
## create a dataframe
a_dataframe <- data.frame(a=c(1,0),b=c(2,0),c=c(3,0),d=c(4,0))
a_dataframe
##   a b c d
## 1 1 2 3 4
## 2 0 0 0 0

## move columns c and d behind column a
move.column.after(df=a_dataframe,col.name=c("c","d"),col.name.after="a")
##   a c d b
## 1 1 3 4 2
## 2 0 0 0 0

}
\seealso{
move.column.to.pos
}
\author{
Pascal Benkert
}
