% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{remove.center.tag}
\alias{remove.center.tag}
\title{Helper function to strip away trailing center tag in parentheses}
\usage{
remove.center.tag(x)
}
\arguments{
\item{x}{center string}
}
\value{
cleaned center string
}
\description{
Helper function to strip away trailing center tag in parentheses
}
\examples{
remove.center.tag("Universitätsspital Basel (SWISS-AF)")
# "Universitätsspital Basel"
remove.center.tag("HUG Genève (SSR)")
# "HUG Genève"
}
\seealso{
mnppid2center
}
